NM: Fix handling of multiple matches
g_string_append_netdef_match() is only being used for blacklisting, so we must
not fail on multiple matches when using networkd. If a MAC is given, prefer
that over names as that should always be unique and avoids running into "NM
does not support name globbing" issues.

For NM output we actually do support matching on name *and* MAC.