Allow drop-ins to update definitions
Change handle_network_type() to update an already existing definition instead
of always creating a new one, so that overriding/merging works properly.

This requires refining the "duplicate ID" check, as this now needs to be
limited to IDs within *one* file (parse_yaml() call). So track the IDs in the
current file in a new hash table "ids_in_file" and only assert that these are
unique.