Add support for VLAN
We need to introduce an UUID for every network definition for this:
In NetworkManager the VLAN device specifies its parent link, but that can only
be an interface name or an UUID. If the parent device uses "match:" we don't
have an interface name.

Only use the UUID in this case for the time being, to retain predictable and
easy to read files for all  other cases.