Add networkd support for wifi
networkd cannot handle wifi natively, so generate a wpa_supplicant
configuration and unit that does the WLAN part and leave the IP part to
networkd.

This supports open and WPA2, and infrastructure and adhoc mode. "match:" is not
supported as the netplan-wpa@.service helper unit is bound to the named network
interface device unit.

tests/integration.py: Bump timeout for settling networkd from 15 to 25 seconds
as the wpa_supplicant part also takes a few seconds.

LP: #1616928