tests: wifi-based match for driver (using mac80211_hwsim)
The only driver matching that I could think would reliably work (using
mac80211_hwsim, which we explicitly install for mocking wifi) during tests
is not supported by systemd-networkd or NetworkManager. This needs more work,
and with it we might be able to add such support, but for now let's skip the
test we just wrote for this and rely on manual testing of the --mapping feature
for searching per-driver. This way the test can be reused once driver matching
works in the backends.

We don't currently have anything else we can rely on to mock network devices,
as veths don't have a device/driver directory under /sys.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>