Fix GLib warning on overwriting errors
When fuzzing, I see warnings like:
GLib-WARNING **: GError set over the top of a previous GError or uninitialized memory.
This indicates a bug in someone's code. You must ensure an error is NULL before it's set.
The overwriting error message was: ...

This happens because the multi-pass parser doesn't clear the errors.
Clear them.

Found with afl-fuzz/FairFuzz.

Signed-off-by: Daniel Axtens <dja@axtens.net>