parser: validate that all netdef elements have unique memory
In preparation for being able to duplicate the entire netdef state
for LP: #1746842, we want to verify that when walking the netdefs,
we never come across two distinct elements that point to the same
memory. This is important because if it is not true, making a deep
copy becomes very complex.

Add the validation routine and fix up anything it reveals.

Signed-off-by: Daniel Axtens <dja@axtens.net>