generate: don't die on first broken YAML file
Fix LP: #1746842: if there is a broken YAML file, it should not
cause all the parsing to fail - just that file should be ignored.
(This is really helpful for not locking people out of cloud VMs!)

To do this, make a copy of the netdefs before parsing each file,
and revert to that copy if parsing the next file does not work.

Also provide --strict mode to keep existing behaviour, which is
especially useful for tests.

[This won't save you if you have an error in the networkd or nm
generation code, but hopefully those errors are fewer.]

Signed-off-by: Daniel Axtens <dja@axtens.net>