docs: clarify that globbing is not supported by NetworkManager
I used the example config in testing and was surprised when it failed
to render because it specified both NetworkManager and globbing.

It's a good example so we don't want to remove the globbing, and we
probably want to keep the NM renderer because of the wifi. So just
document that it's an example only and not expected to work out of
the box.

Signed-off-by: Daniel Axtens <dja@axtens.net>