docs: document 'resend-igmp'
I did an exhaustive enumeration of accepted options and this was
the last undocumented one.

Documentation is informed by https://www.kernel.org/doc/Documentation/networking/bonding.txt
I didn't want to include it verbatim as it will be GPLv2 or GPLv2+,
and I'm not sure we can relicense that as GPLv3 (only) which
is what netplan is under.

Signed-off-by: Daniel Axtens <dja@axtens.net>