Cope with multiple nexthops for default gateway in tests
The test bed might be set up in such a way that there are multiple
next hops for the default gateway, and there tends to be variation
in how NM and networkd set up the default gateway on IPv6. Cope
with these differences by just looking for 'via $gateway'. It's
sufficient to tell that the configuration we passed to the backend
has been applied.