Move replug blacklisted devices into a config file
Currently, devices that are blacklisted for replug are hardcoded
in apply.py. This is going to make support harder as people will
have to patch a python file to add another blacklist entry.

Move it to a yaml file - /usr/share/netplan/blacklist.yaml.

Also allow people to supply /etc/netplan/blacklist so they can
test blacklisting of a device or to apply a workaround while
waiting for the package to go through the SRU process.

TODO: install the /etc/ file.

Signed-off-by: Daniel Axtens <dja@axtens.net>