Support blacklisting replug devices in config files
Currently, devices that are blacklisted for replug are hardcoded
in apply.py. This is going to make support harder as people will
have to patch a python file to add another blacklist entry.

Introduce `replug:` as a mapping at the top level that contains:

 `disable_all_replug`: a kill switch for replugging

 `blacklist`: a sequence of mappings that describes devices to be
              blacklisted

The contents of the current list is moved to blacklist.yaml which
is installed in /lib/netplan/00-netplan-replug-blacklist.yaml

Signed-off-by: Daniel Axtens <dja@axtens.net>