accept-ra: do not enable by default, default to unset
Do not always emit an IPv6AcceptRA value into network configurations.
Systemd-networkd defaults to kernel value as long as the value stays
unset in the configuration.  When IPv6AcceptRA is enabled in systemd
this makes networkd wait 10 seconds for potential RAs to arrive.  This
delays boot by 10 seconds always whether or not the network has an
IPV6 Router running.  Leaving things unsets keeps existing behavior
of Ubuntu systems where RAs are accepted and processed at the time
they are received and boot is not delayed by waiting for an RA to arrive.

If users want to forcibly enable or disable, they may do so by including
a value for accept-ra in their yaml configuraion.

LP: #1732002