apply: Make it possible to run daemon restarts synchronously
For the 'netplan try' command; always run 'apply' synchronously, as otherwise
we might spend some of the wait time for input waiting for the backend to
finish restarting, or worse, have the user acknowledge changes that are not
actually fully applied yet.

There is still a question of network convergence time for STP and things like
that, but at least we know the system has already tried to make the changes,
and we typically wait long enough for input to ensure things don't timeout
before the user could acknowledge the changes.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>