tests: add script to validate that all YAML keys are documented
In f18aa632469e ("docs: document 'optional'") and
29a22cbac093 ("docs: document 'resend-igmp'") I tried to do an
exhaustive manual enumeration of YAML keys and make sure they were
all documented.

It seems I missed 'mtu', and since then errors were also added in
e2c28b724533 ("doc: update doc for per-interface routes: and routing-policy:")
with fwmark vs mark.

Create a script that attempts to parse parse.c and check if any YAML
key that is defined is undocumented. It's not perfectly precise as
it uses regexes to parse C, but it seems to work well - it picked
up the (fw)mark issue and I also tested it against my fixes to
optional, resend-igmp, and mtu and it would have picked them up too.

Make it part of make check.

Signed-off-by: Daniel Axtens <dja@axtens.net>