docs: create man pages for netplan apply, generate and try
Now you can do "man netplan apply" (yes, with the spaces!) and get
information specific to the commands. It also works for generate and
try. Nothing for 'ip leases' just yet.

Designed to complement netplan(5), the pages specifically describe
the functionality of the commands and refer the reader to netplan(5)
for the YAML format.

Signed-off-by: Daniel Axtens <dja@axtens.net>