Spell Gratuitous ARP correctly and make it work
LP: #1756701

Throughout the codebase, gratuitous has an extra 'i' (gratuit_i_ous).

This includes the output for networkd. This misspelling causes the
value to have no effect - regardless of the setting, I only see 1
gratuitous ARP in active-backup mode.

It also causes the following message to be printed to syslog:

/run/systemd/network/10-netplan-bond0.netdev:9: Unknown lvalue 'GratuitiousARP' in section 'Bond'

Changing the networkd code to spit out the correct spelling causes
the correct behaviour.

The NM parameter is num_grat_arp, which is not affected by the broader
misspelling. However, NM is affected by a different bug which fixes the
number of gratuitous arps to 1. This has been fixed upstream but not
backported yet:
https://github.com/NetworkManager/NetworkManager/commit/42b0bef33c77a0921590b2697f077e8ea7805166

The workaround is trivial - just set num_unsol_na to the same value:
they map to the same kernel parameter internally anyway. So implement
the workaround.

 - Use the correct spelling in the code
 - Accept both new and old spellings in config files
 - Update the documents and tests appropriately

Signed-off-by: Daniel Axtens <dja@axtens.net>