Add ability to set interface route metrics
LP: #1771834

Currently, we set a route metric of 600 for routes on a wifi device
and 100 on other devices. We allow static routes to set their own
metric, but have no way to set the metric for routes that come from
DHCP, or the route to the subnet. Add this ability through a 'metric'
key on an interface.

This is particularly useful for systems with multiple DHCP interfaces
where one should be preferred.

Add a 'metric' key to interfaces, plumb it through, document it,
add tests.

Signed-off-by: Daniel Axtens <dja@axtens.net>