Disable 10-globally-managed-devices.conf if any device uses NM
LP: #1753860

I found that specifying NM as the renderer for a single ethernet
device didn't work - 10-globally-managed-devices.conf still stops
NM from managing it.

Disable that file if any interface uses NM.

I am a little concerned that this might lead to different behaviour
if you hotplug a device when any device is using NM - it will be
managed by NM. This already happens if NM is the global renderer, so
I don't think it's a big issue, but something to be aware of.

Signed-off-by: Daniel Axtens <dja@axtens.net>