apply: skip non-physical and bond interfaces when applying name changes (LP: #1802322)
set-name applies only to physical interfaces.  physical interfaces which
are part of a bond or bridge may undergo a mac address change due to how
composition works.  Netplan cannot know if the current mac of the interface
is in fact the correct mac address and should instead skip renaming interfaces
that  are part of a composite device.

Based on a patch by Ryan Harper, with added tests and fixups.