parser: fix explicit backend assignment due to file merging
In the past, backend was left as "NONE" until a post-processing pass where
we'd set it to the default for the type, usually 'networkd'. However, this
breaks validating netdefs prior to that postprocessing, and it would be wrong
to wait until post-processing to do that validation.

Instead, explicitly set the backend to the default by type, and adjust our
merging test to account for that. This effectively changes the behavior such
that a change of renderer in a later file will not apply, but it's less
confusing this way: a "global" backend only affects the file where it's
defined.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>