tunnels: rework key validation for clarity and correctness
Turns out I initially got things reversed in the tests, where NM was testing
for keys in VTI* tunnels and networkd for keys in GRE tunnels. It's the other
way around.

Debugging this showed how the initial code for validating these was confusing
in validate_tunnel(), so I reworked it to use nested switches, which makes it
more verbose so hopefully easier to follow.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>