apply: re-start services also when unconfiguring
We were not re-starting networkd/NetworkManager in the case we moved
from a configuration that created netplan files in /run to one that did
not need files there. Now we make sure we re-start the services also in
that case, by checking if there were configuration files before the
generation step. Fixes LP: #1811868.