Skip some tests on codecov.io
GLib apparently has some differences in the order in which members are
iterated in a GHashTable, which leads to things outputting from a GHashTable
to turn up in different order depending on the GLib version used to build
the generator. Skip these tests on codecov.io: we do have a pretty good test
coverage anyway.

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>