parser: refactor / factor out validation and error reporting
Move validation and error reporting to their own files; also rework
the validation passes slightly to validate as much as we can of the
netplan grammar in parser passes (so we get context info for errors),
and leave the backend rules sanity checking only to after the parsing.

This way we'll get the error context, unless things are backend-specific
rules for which it's just hard to get the info (and not as helpful).