add dbus based "netplan apply" interface
This PR adds a new io.netplan.Netplan dbus activatable service
that can be used to Apply a netplan configuration.

The main use-case for this is Ubuntu Core where we want to allow
snaps to control the netplan network configuration via an interface.

However we do not want to give access to systemctl to restart
systemd-networkd or network-manager. The dbus service will run
outside of the confinement will archive this goal.

The interface is probably useful beyond Ubuntu Core.