parser: add bare-bones parsing logic for NetworkManager-specific settings
This will allow recording the UUID, name, etc. for a connection from NM's
viewpoint, such that on connection updates in UI or API in NetworkManager,
these changes can be read back to ensure the NMStorage is valid (ie. writes
happen correctly and there isn't corruption or another process making changes)

Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>