Handled wired wpa_supplicant driver automatically
Currently netplan can't handle authentication for wired connections
because it doesn't force the loading of the wired driver.

This is based on @fluffy's work in #80 but re-working it to drop
the existing pre-written unitfiles and instead to generate our own,
as well as automatically determine when to load the wired driver.

Signed-off-by: Phil Dibowitz <phil@ipom.com>