SR-IOV support for explicitly defining the VF count (#130)
This is a possible feature request to the just-landed SR-IOV support in netplan. This PR introduces an optional 'virtual-function-count:' parameter that can be defined for physical functions to force the allocation of a given number of VFs, regardless of how many are actually used in the netplan config. There are of course safety checks to ensure that we can't request less VFs than actually needed in netplan.

This feature request came from the OpenStack team. We did consider such a thing originally, but then decided that it's better if we let netplan handle it. This PR adds this as an option (not recommended for everyday usage tho).

Co-authored-by: Łukasz 'sil2100' Zemczak <lukasz.zemczak@canonical.com>