Initial first-try framework for OpenVSwitch schema
This basically adds support for generating systemd unit files that would then later configure the OVS bits for devices that require it. The only schema piece supported right now is 'other-config' and 'external-ids', which were quite easy to do as they directly pass-through to OVS. Global setting not yet supported.

There is a basic OVS test to confirm that the parser and generator changes work.