Fix process_link_changes handling 'up' interfaces
b7f1d9b04212 refactored process_link_changes with helper methods to get
the interface driver name and MAC address. This new code introduced a
regression where it's possible for an interface in the up state to be
included in the changelist by its MAC address.

This patch restores the previous behaviour, skipping interfaces that
don't have driver_name set.

Fixes: https://bugs.launchpad.net/bugs/1875411