Fix autopkgtest on arm64 with NM 1.24 (#146)
On arm64 we had a test failure in tests/integration/ethernets.py -> test_manual_addresses, which was caused by the fact that netplan apply was disconnecting ALL available network interfaces via nmcli, in order to reload the new netplan config. This way the IP of the 2nd test DHCP server (dnsmasq on veth43) was lost and did not provide correct responses for the rest of the test.

It is not necessary to take all interfaces down, only the ones which are (or were) configured via netplan. This was fixed in the netplan apply command.

Note: This should make its way into the Groovy package as a distro-patch, to avoid future autopkgtest failures.

Commits:
* tests: fix ethernets/test_manual_addresses on arm64
* netplan:apply: disconnect/re-configure only netplan-NM interfaces
* tests:ethernets: remove deprecated fix, was fixed via 'netplan apply'
* tests: add unit-test for python utils
* utils: get rid of default glob in nm_interfaces