Work on a better first-boot experience with SR-IOV (#141)
Currently there's no real first boot experience in SR-IOV. I propose adding some simple udev rules so that SR-IOV initialization can be performed on first boot and whenever a new SR-IOV-enabled device appears. The rule then calls netplan apply SR-IOV-only part, which is fine to be executed multiple times. We could add more fine grained control and only perform initialization of the device that's appearing, but that would require additional code changes - and it's fine as is like it is right now.

Co-authored-by: Łukasz 'sil2100' Zemczak <lukasz.zemczak@canonical.com>