Implement just-in-time behaviour for generate (#162)
If system is initializing (basic.target not reached yet), and netplan generate is called ensure that any netplan generated service units are added to the initial boot transaction.

This should resolve cloud-init first-time booting with systemd-networkd disabled, or with requirement to add wlan/wifi units.

Commits:
* generate: implement just-in-time behaviour of generate

If system is initializing (basic.target not reached yet), and `netplan
generate` is called ensure that any netplan generated service units
are added to the initial boot transaction.

This should resolve cloud-init first-time booting with
systemd-networkd disabled, or with requirement to add wlan/wifi units.

* Add integration tests for cloud-init OVS/WPA first-boot

* generate: coverage 100%, by excluding special parts, covered by the integration test

* generate: jit starting of units only as root

Units shall only be started JIT during early boot, by the system user
(root). If a normal user calls 'netplan generate' it shall not start the
units. Avoid asking for a password if a user (or test) executes this
command and rather fail with missing authentication.

* Add documentation and feature flag

* generate: no jit if network.target already started

Do not try to enqueue new network related netplan-*.service units, if
network.target was already started.

Co-authored-by: Lukas Märdian <lukas.maerdian@canonical.com>