Implement DBus try-commit handling (#134)
Implement the Try(uint32 timeoutInSeconds) and Cancel() command inside netplan-dbus and adopt the Apply() command according to spec. "Try(...)" calls netplan try, keeping stdin open. The "Apply()" command uses the SIGUSR1 signal, to send a confirmation to the netplan try process to confirm/commit the settings. The "Cancel()" command uses the SIGINT signal, to send a rejection to the netplan try child process.
The child process is being monitored, using events on the sd-event mainloop.

Commits:
* dbus: initial draft for try-commit handling
* dbus: improve try tests
* dbus: make linter happy
* cli:try: accept config on USR1 signal
* Revert changes to netplan/terminal.py
* Revert "Revert changes to netplan/terminal.py"
* dbus: use SIGINT/SIGUSR1 to cancel/confirm TRY
* dbus:try: check child status via waitpid/WNOHANG
* tests:dbus: adopt try test case
* terminal: make coverage happy
* dbus: integrate with sd-event mainloop
* dbus: avoid global *bus variable
* dbus: avoid global vars, use signal handlers, pass userdata
* dbus: cleanup
* dbus: move Confirm() into Apply(), as defined by the spec
* dbus: cleanup
* tests: try to improve codecov coverage
* Revert "tests: try to improve codecov coverage"
* dbus: cleanup sd_event
* dbus: remove unneeded safeguards for easier testability