Add support for the activation-mode parameter (LP: #1664844) (#171)
Per an existing specification, add the activation-mode parameter which currently supports the 'off' and 'manual' modes. This allows administrators either force a given interface to be always in the down state ('off', only for networkd right now) or allow the user to manually manage the state ('manual'). This means that a 'manual' interface will only be up/down if the user does that explicitly.

The main target is the networkd backend, but we try supporting at least the manual setting for NetworkManager. The networkd support for this has now been merged into v248 (which is still not released as of this writing).

https://bugs.launchpad.net/netplan/+bug/1664844

COMMITS:
* Add support for the activation-mode parameter, used for specifying the interface administration policy.
* Add feature flag, clarify some documentation/comments.
* Update src/nm.c
Co-authored-by: Lukas Märdian <lukas.maerdian@canonical.com>
* Update doc/netplan.md
Co-authored-by: Lukas Märdian <lukas.maerdian@canonical.com>
* Update src/networkd.c
Co-authored-by: Lukas Märdian <lukas.maerdian@canonical.com>
* Move the new activation_mode struct member to the end of the struct and bump the netplan minimum version for the feature.
* Update src/nm.c
Co-authored-by: Lukas Märdian <slyon@ubuntu.com>