Improve integration tests interface-ready waiting logic (LP: #1922126) (#204)
This is a big refactoring of the integration test suite, which should make it much faster and more stable in different testing environments (e.g. armhf LXD containers).

The main change is in tests/integration/base.py and changes the waiting logic to decide when an given interface is readily configured. Previously, we were relying on /lib/systemd/systemd-networkd-wait-online to tell us magically when the networking is fully configured. This does not properly work for interfaces controlled by NetworkManager and is only watching the "general" networking state (i.e. whenever the first interface is up and running).

Waiting only on an arbitrary "first" network interface to be ready leads to races as we use asserts on many individual interfaces, which might not be ready at that time. So with this change the generate_and_settle(self, wait_interfaces=None): method got a new wait_interfaces argument, where the developer needs to specify each interface to be tested in a given test case. Furthermore, the developer can wait for a specified state (i.e. DHCP4/6 address assigned), to account for waiting on asynchronous background tasks, e.g.:

self.generate_and_settle([self.dev_e_client, 'eth43', self.state_dhcp4('mybond')])

All existing integration tests have been adopted to specify their interfaces-under-test in generate_and_settle(). Additionally, some logging was added to always print the name of the currently waited on interface and one dot per second (up to 60s), to see how long it takes to become ready (or times out), especially on slower architectures.

COMMITS:
* tests:base: improve waiting logic (per interface)
* tests: adopt new waiting logic and copyright
* tests:base: don't ignore asserts
* tests:routing: fix route_on_link
* tests:base: improve DHCP server reset
* tests: wait for DHCP IPs
* tests:wifi:scenarios: fixes for Hotspot and Bridge