parse: handle_routes: clean up the route on error
This patch reworks the single route processing to only append the parsed
route to the interface when it has been fully validated, and use a
single error path otherwise. This allows us to locally control the
cleanup and freeing of the allocated route.

The previous version could indeed leak if process_mapping returned an
error (and that's assuming that the interface cleanup process would free
up the routes already assigned, which I haven't checked).