parse: routes: avoid creating dangling pointers on error
Parsing a yaml route looking like this:

```yaml
routes:
- to: 162.168.0.0/16
- to: invalid value
```

would result in having a dangling pointer in cur_route->to, as its
content would have been freed when starting to parse the invalid value,
without the pointer itself having been cleared. By deferring the free() to
right before the pointer update, we ensure that (barring interrupts or
multithreading, neither of which are supported AFAIK) the pointer remain
valid.

Note that to be thread-safe, we'd need to do an atomic swap and only
then free the variable, but that's out of scope ATM.