Gateway fields deprecation and default routing support (FR-728) (LP: #1756590) (#216)
This patchset deprecates the gateway4/6 fields, advising the user to use a proper route instead. The rationale is that it groups all routing matters in the same place, making things more consistent. Along with this deprecation comes the support for the "default" keyword in the route to field, and an expansion of the gateway validation step to encompass all default routes.

Examples
```
network:
  ethernets:
    en1:
      addresses:
      - 1.2.3.4/8
      - fc00:123:4567:89ab:cdef::1234/64
      routes:
      - to: default
        via: 1.1.1.1
      - to: default
        via: "fc00:123:4567:89cb:cdef::1"
```

Addresses LP: #1756590

COMMITS:
* tests: split gateway4/6 into their own tests
* doc: add a paragraph regarding default routes
* netplan: deprecate gateway4/6
* tests: gw: add a multipass check
* parse: accept default routes as valid input
* validation: look at all routes for default route validation
This makes the default route validation much more comprehensive, looking
at all the routes. The underlying code is a bit more involved as a
result, though.
* networkd: support default route specification
* nm: support default route specification
* Rewording of error messages for consistency
Co-authored-by: Lukas Märdian <slyon@ubuntu.com>
* Whitespace fixes in string literals
Co-authored-by: Lukas Märdian <slyon@ubuntu.com>
* tests: add an integration test for default routes
* tests:integration:routing: not using DHCP in routes_default
Co-authored-by: Lukas Märdian <slyon@ubuntu.com>